[FIX] CSS changes and folder restructuring
