+++ IKS/uber/src/app/app.routes.ts
@@ -6,7 +6,7 @@ import { HomeComponent } from './layout/home/home.component';
 import { AccountComponent } from './layout/account/account.component';
 import { RideHistoryComponent } from './driver/ride-history/ride-history.component';
 import { RequestsComponent } from './layout/requests/requests.component';
-import { TrackingRouteComponent } from './tracking-route/tracking-route.component';
+import { TrackingRouteComponent } from './layout/tracking-route/tracking-route.component';
 
 
 export const routes: Routes = [
