+++ IKS/uber/src/app/layout/tracking-route/tracking-route.component.css
@@ -6,7 +6,7 @@ div{
 }
 
 #container{
-    padding-top: 10%;
+    padding-top: 5%;
 }
 
 p{
